make form component
